QWizard: Support fractional DPR on Windows
We were storing the DPR as int; change to qreal.

Task-number: QTBUG-114175
(cherry picked from commit c0e48ed645f68e714e982064540d710676a0738b)

+++ src/widgets/dialogs/qwizard_win.cpp
@@ -69,7 +69,7 @@ Q_DECLARE_METATYPE(QMargins)
 QT_BEGIN_NAMESPACE
 
 int QVistaHelper::instanceCount = 0;
-int QVistaHelper::m_devicePixelRatio = 1;
+qreal QVistaHelper::m_devicePixelRatio = 1.0;
 QVistaHelper::VistaState QVistaHelper::cachedVistaState = QVistaHelper::Dirty;
 
 /******************************************************************************
@@ -119,7 +119,7 @@ void QVistaBackButton::paintEvent(QPaintEvent *)
     RECT clipRect;
     int xoffset = origin.x() + QWidget::mapToParent(r.topLeft()).x() - 1;
     int yoffset = origin.y() + QWidget::mapToParent(r.topLeft()).y() - 1;
-    const int dpr = devicePixelRatio();
+    const qreal dpr = devicePixelRatio();
     const QRect rDp = QRect(r.topLeft() * dpr, r.size() * dpr);
     const int xoffsetDp = xoffset * dpr;
     const int yoffsetDp = yoffset * dpr;
